Sabado 08/06/2024

+++ src/main/java/com/dam/Proyecto/dao/PluviometroDao.java
@@ -14,7 +14,7 @@ public interface PluviometroDao {
 
     void registrar(Pluviometro pluviometro);
 
-    Pluviometro getPluviometro(Long id);
+    Pluviometro getPluviometro(int id);
 
     Long contarPluviometros();
 }
